+++ src/ctrls/edit.rs
@@ -537,7 +537,7 @@ impl CEdit {
     }
 
     #[inline(always)]
-    pub fn GetDlgItem2 (&self,nID:c_int) -> CWindow {
+    pub fn GetDlgItem2 (&self,nID:WORD) -> CWindow {
         self.cwin.GetDlgItem2(nID)
     }
