+++ src/ctrls/static_ctrl.rs
@@ -175,7 +175,7 @@ impl CStatic {
     }
 
     #[inline(always)]
-    pub fn GetDlgItem2 (&self,nID:c_int) -> CWindow {
+    pub fn GetDlgItem2 (&self,nID:WORD) -> CWindow {
         self.cwin.GetDlgItem2(nID)
     }
